+++ src/transports/expressBasedHttpServer.ts
@@ -1,0 +1,99 @@
+import express from "express";
+import type http from "http";
+import { LogId } from "../common/logging/loggingDefinitions.js";
+import type { LoggerBase } from "../lib.js";
+
+export type ExpressConfig = {
+    port: number;
+    hostname: string;
+};
+
+/** @internal */
+export abstract class ExpressBasedHttpServer {
+    protected httpServer: http.Server | undefined;
+    protected app: express.Express;
+
+    protected readonly logger: LoggerBase;
+    protected readonly logContext: string;
+
+    protected readonly expressConfig: ExpressConfig;
+
+    constructor(config: { logger: LoggerBase; logContext: string } & ExpressConfig) {
+        this.app = express();
+        this.app.enable("trust proxy"); // needed for reverse proxy support
+        this.expressConfig = { port: config.port, hostname: config.hostname };
+
+        this.logger = config.logger;
+        this.logContext = config.logContext;
+    }
+
+    public get serverAddress(): string {
+        const result = this.httpServer?.address();
+        if (typeof result === "string") {
+            return result;
+        }
+        if (typeof result === "object" && result) {
+            return `http://${result.address}:${result.port}`;
+        }
+
+        throw new Error("Server is not started yet");
+    }
+
+    protected abstract setupRoutes(): Promise<void>;
+
+    public async start(): Promise<void> {
+        await this.setupRoutes();
+
+        const { port, hostname } = this.expressConfig;
+
+        this.httpServer = await new Promise<http.Server>((resolve, reject) => {
+            const result = this.app.listen(port, hostname, (err?: Error) => {
+                if (err) {
+                    reject(err);
+                } else {
+                    resolve(result);
+                }
+            });
+        });
+
+        this.logger.info({
+            message: `Http server started on address: ${this.serverAddress}`,
+            context: this.logContext,
+            noRedaction: true,
+            id: LogId.httpServerStarted,
+        });
+    }
+
+    public async stop(): Promise<void> {
+        if (this.httpServer) {
+            this.logger.info({
+                message: "Stopping server...",
+                context: this.logContext,
+                id: LogId.httpServerStopping,
+            });
+
+            const server = this.httpServer;
+
+            await new Promise((resolve, reject) => {
+                server.close((err?: Error) => {
+                    if (err) {
+                        reject(err);
+                    } else {
+                        resolve(undefined);
+                    }
+                });
+            });
+            this.logger.info({
+                message: "Server stopped",
+                context: this.logContext,
+                id: LogId.httpServerStopped,
+            });
+        } else {
+            this.logger.info({
+                message: "Server is not running",
+                context: this.logContext,
+                id: LogId.httpServerStopped,
+            });
+        }
+    }
+}
